MOTOR-40 Move test-environment discovery into test_environment.py.
