+++ test/test_motor_client.py
@@ -32,7 +32,8 @@ from tornado.testing import gen_test, netutil
 import motor
 import motor.core
 import test
-from test import host, port, assert_raises, MotorTest, SkipTest
+from test import assert_raises, MotorTest, SkipTest
+from test.test_environment import host, port, env
 from test.motor_client_test_generic import MotorClientTestMixin
 from test.utils import remove_all_users, delay
