+++ test/test_motor_replica_set.py
@@ -26,8 +26,9 @@ from tornado.testing import gen_test
 import motor
 import motor.core
 import test
-from test import host, port, MotorReplicaSetTestBase, assert_raises, MotorTest
+from test import MotorReplicaSetTestBase, assert_raises, MotorTest
 from test import SkipTest
+from test.test_environment import port, host, env
 from test.motor_client_test_generic import MotorClientTestMixin
